+++ flutter_base_loader/lib/plugins/loading.dart
@@ -1,26 +1,31 @@
 import 'dart:async';
 
+import 'package:flutter/material.dart';
+
 import './base_info.dart';
 
 class BaseLoading {
   final Completer<void> _com = Completer<void>();
-  final Completer<void> _comLoadingPage = Completer<void>();
+  final Completer<BuildContext> _comLoadingPage = Completer<BuildContext>();
   factory BaseLoading() => _instance;
   static final BaseLoading _instance = BaseLoading._internal();
   late Future<void> _future;
-  late Future<void> _futureReady;
+  late Future<BuildContext> _futureReady;
 
   BaseLoading._internal() {
-    _futureReady = _comLoadingPage.future.then((_) => BaseInfo.init());
-    _future = _futureReady.then((value) => _com.future);
+    _futureReady = _comLoadingPage.future.then<BuildContext>((ctx) {
+      BaseInfo.init();
+      return ctx;
+    });
+    _future = _futureReady.then((_) => _com.future);
   }
   static BaseLoading get instance => _instance;
 
   Future<void> get future => _future;
-  Future<void> get onReady => _futureReady;
+  Future<BuildContext> get onReady => _futureReady;
 
-  void pageReady() {
-    if (!_comLoadingPage.isCompleted) _comLoadingPage.complete();
+  void pageReady(BuildContext context) {
+    if (!_comLoadingPage.isCompleted) _comLoadingPage.complete(context);
   }
 
   void complete() {